requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.7
pymongo==4.6.3
python-dotenv==1.0.1
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
